+++ lib/std/enumutils.nim
@@ -86,48 +86,26 @@ iterator items*[T: HoleyEnum](E: typedesc[T]): T =
     assert B[float].toSeq == [B[float].b0, B[float].b1]
   for a in enumFullRange(E): yield a
 
-func span(T: typedesc[HoleyEnum]): int =
-  (T.high.ord - T.low.ord) + 1
-
-const invalidSlot = uint8.high
-
-proc genLookup[T: typedesc[HoleyEnum]](_: T): auto =
-  const n = span(T)
-  var ret: array[n, uint8]
-  var i = 0
-  assert n <= invalidSlot.int
-  for ai in mitems(ret): ai = invalidSlot
-  for ai in items(T):
-    ret[ai.ord - T.low.ord] = uint8(i)
-    inc(i)
-  return ret
+import std/private/lookuptables
 
 func symbolRankImpl[T](a: T): int {.inline.} =
-  const n = T.span
-  const thres = 255 # must be <= `invalidSlot`, but this should be tuned.
-  when n <= thres:
-    const lookup = genLookup(T)
-    let lookup2 {.global.} = lookup # xxx improve pending https://github.com/timotheecour/Nim/issues/553
-    #[
-    This could be optimized using a hash adapted to `T` (possible since it's known at CT)
-    to get better key distribution before indexing into the lookup table table.
-    ]#
-    {.noSideEffect.}: # because it's immutable
-      let ret = lookup2[ord(a) - T.low.ord]
-    if ret != invalidSlot: return ret.int
-  else:
-    var i = 0
-    # we could also generate a case statement as optimization
-    for ai in items(T):
-      if ai == a: return i
-      inc(i)
+  const lut = (proc(): auto = # pending https://github.com/nim-lang/RFCs/issues/276
+    var a: seq[int]
+    for ai in items(T): a.add ai.ord
+    a.initLookupTable)()
+  let lut2 {.global.} = lut
+    # xxx improve pending https://github.com/timotheecour/Nim/issues/553
+  {.noSideEffect.}: # because it's immutable
+    let ret = lut2.lookup(a.ord)
+  if ret != -1: return ret
   raise newException(IndexDefect, $ord(a) & " invalid for " & $T)
 
 template symbolRank*[T: enum](a: T): int =
-  ## Returns the index in which `a` is listed in `T`.
-  ##
-  ## The cost for a `HoleyEnum` is implementation defined, currently optimized
-  ## for small enums, otherwise is `O(T.enumLen)`.
+  ## Returns the index in which `a` is listed in `T` with `O(1)` cost even
+  ## for `HoleyEnum`.
+  # The `O(1)` cost for `HoleyEnum` should hold except for adverserially designed
+  # holey enums; if that becomes a problem we could fix it by optimizing over
+  # `lookuptables.pseudoRandomMixing` since `T` is known at CT.
   runnableExamples:
     type
       A = enum a0 = -3, a1 = 10, a2, a3 = (20, "f3Alt") # HoleyEnum
@@ -145,8 +123,6 @@ template symbolRank*[T: enum](a: T): int =
 
 func symbolName*[T: enum](a: T): string =
   ## Returns the symbol name of an enum.
-  ##
-  ## This uses `symbolRank`.
   runnableExamples:
     type B = enum
       b0 = (10, "kb0")